[enh] using sqlite3 datatype REAL for gps coords

+++ src/initialization.sql
@@ -3,8 +3,8 @@ CREATE TABLE Stations(
     name        VARCHAR(64),
     payTerminal BOOLEAN,
     capacity    SMALLINT UNSIGNED,
-    gpsx        DECIMAL(9,6),
-    gpsy        DECIMAL(9,6),
+    gpsx        REAL,
+    gpsy        REAL,
 
     PRIMARY KEY (id)
 );
